Fix MaxBeatStrength type
Float not int :grimace:

+++ Source/LETSGO/MusicEngine/MusicComposition/Strategy_CreateMotif.cpp
@@ -30,7 +30,7 @@ struct FCreateMotifData
 	int Beats;
 
 	UPROPERTY()
-	int MaxBeatStrength;
+	float MaxBeatStrength;
 
 	UPROPERTY()
 	TMap<int, float> BeatStrength;
